+++ src/search/database/hash_semi_join.cc
@@ -7,8 +7,8 @@ using namespace std;
 
 size_t hash_semi_join(Table &t1, Table &t2) {
     vector<pair<int, int>> matches;
-    for (int i = 0; i < t1.tuple_index.size(); ++i) {
-        for (int j = 0; j < t2.tuple_index.size(); ++j) {
+    for (size_t i = 0; i < t1.tuple_index.size(); ++i) {
+        for (size_t j = 0; j < t2.tuple_index.size(); ++j) {
             if (t1.tuple_index[i] == t2.tuple_index[j])
                 matches.emplace_back(i, j);
         }
@@ -29,7 +29,7 @@ size_t hash_semi_join(Table &t1, Table &t2) {
         // Build phase
         for (const vector<int> &tuple : t2.tuples) {
             vector<int> key(matches.size());
-            for (int i = 0; i < matches.size(); i++) {
+            for (size_t i = 0; i < matches.size(); i++) {
                 key[i] = tuple[matches[i].second];
             }
             hash_join_map[key].insert(tuple);
@@ -37,7 +37,7 @@ size_t hash_semi_join(Table &t1, Table &t2) {
 
         for (const vector<int> &tuple : t1.tuples) {
             vector<int> key(matches.size());
-            for (int i = 0; i < matches.size(); i++) {
+            for (size_t i = 0; i < matches.size(); i++) {
                 key[i] = tuple[matches[i].first];
             }
             if (hash_join_map.count(key) > 0) {
